[~] separated user routes from auth.
