+++ effectiveOfficeBackend/src/main/kotlin/office/effective/features/user/routing/AuthRouting.kt
@@ -40,24 +40,4 @@ fun Route.authRoutingFun() {
         }
     }
 
-    get("/usersTest/{email}") {
-        val repo: UserRepository = GlobalContext.get().get()
-        val model =
-            repo.findByEmail((call.parameters["email"] ?: call.response.status(HttpStatusCode.BadRequest)) as String)
-        val converterDTO = UserDTOModelConverter()
-        call.respond(converterDTO.modelToDTO(model))
-    }
-    get("/users") {
-        var tagStr = call.request.queryParameters["tag"] ?: call.response.status(HttpStatusCode.BadRequest)
-        val tokenStr = call.request.header("id_token") ?: call.response.status(HttpStatusCode.Forbidden)
-        val users: Set<UserDTO>? = service.getUsersByTag(tagStr as String, tokenStr as String)
-        call.respond(users ?: "no such users")
-    }
-    get("/users/{user_id}") {
-        val userId = call.parameters["user_id"] ?: call.response.status(HttpStatusCode.BadRequest)
-        val tokenStr = call.request.header("id_token") ?: call.response.status(HttpStatusCode.Forbidden)
-        val user = service.getUserById(userId as String, tokenStr as String)
-        call.respond(user ?: "No such user. Suggestion: bad id")
-    }
-
 }
