+++ effectiveOfficeBackend/src/main/kotlin/office/effective/plugins/Routing.kt
@@ -6,6 +6,7 @@ import io.ktor.server.application.*
 import io.ktor.server.plugins.statuspages.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
+import office.effective.features.user.routing.userRouting
 import office.effective.features.workspace.routes.workspaceRouting
 
 fun Application.configureRouting() {
@@ -15,6 +16,7 @@ fun Application.configureRouting() {
         }
         authRoutingFun()
         workspaceRouting()
+        userRouting()
     }
 
 }
